refactor(proxy): extract DDL handlers
Move table lifecycle, table creation, and in-place ALTER backfill handling into ddl.cc.
Register ddl.cc in the proxy build and keep handlerton/sysvar access behind handler helpers.
